+++ app/src/main/java/com/snoopgame/devices/objectsForJSON/Employee.java
@@ -1,12 +1,12 @@
 package com.snoopgame.devices.objectsForJSON;
 
 public class Employee {
-    private long id;
+    private int id;
     private String name;
     private String surname;
     private String middleName;
 
-    public Employee(long id, String name, String surname, String middleName) {
+    public Employee(int id, String name, String surname, String middleName) {
         this.id = id;
         this.name = name;
         this.surname = surname;
@@ -42,7 +42,7 @@ public class Employee {
         this.middleName = middleName;
     }
 
-    public long getId() {
+    public int getId() {
 
         return id;
     }
